# api/CustomerAPI/Models/AddressDto.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace CustomerAPI.Models
{
    public class AddressDto
    {
        public int ZipCode { get; set; }
        public string Country { get; set; }
    }
}